moved all customer info related dart files
to a single folder (para mas sayun)
+ it should be working pls
